pyproject.toml
[project]
name = "repyphirenderer"
version = "0.1.0"
description = "基于 Python 的 Phigros 谱面播放 / 渲染器"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "moderngl>=5.12.0",
    "numpy>=2.4.0",
    "pygame-ce>=2.5.6",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
default = true
